+++ gecode/support/bitset-offset.hpp
@@ -55,7 +55,7 @@ namespace Gecode { namespace Support {
     /// Offset
     int _offset;
   public:
-    /// Bit set with space for \a s bits with offset of \o
+    /// Bit set with space for \a s bits with offset of \a o
     BitSetOffset(A& a, unsigned int s, int o);
     /// Copy bit set \a bs
     BitSetOffset(A& a, const BitSetOffset& bs);
@@ -103,19 +103,28 @@ namespace Gecode { namespace Support {
 
   template<class A>
   forceinline bool
-  BitSetOffset<A>::get(int i) const { return BitSetBase::get(i-_offset); }
+  BitSetOffset<A>::get(int i) const { 
+    return BitSetBase::get(static_cast<unsigned int>(i-_offset));
+  }
 
   template<class A>
   forceinline void
-  BitSetOffset<A>::set(int i) { BitSetBase::set(i-_offset); }
+  BitSetOffset<A>::set(int i) {
+    BitSetBase::set(static_cast<unsigned int>(i-_offset));
+  }
 
   template<class A>
   forceinline void
-  BitSetOffset<A>::clear(int i) { BitSetBase::clear(i-_offset); }
+  BitSetOffset<A>::clear(int i) {
+    BitSetBase::clear(static_cast<unsigned int>(i-_offset));
+  }
 
   template<class A>
   forceinline int
-  BitSetOffset<A>::next(int i) const { return _offset + BitSetBase::next(i-_offset); }
+  BitSetOffset<A>::next(int i) const {
+    return _offset + static_cast<int>
+      (BitSetBase::next(static_cast<unsigned int>(i-_offset)));
+  }
 
   template<class A>
   void
@@ -126,15 +135,22 @@ namespace Gecode { namespace Support {
 
   template<class A>
   forceinline int
-  BitSetOffset<A>::offset(void) const { return _offset; }
+  BitSetOffset<A>::offset(void) const {
+    return _offset;
+  }
 
   template<class A>
   forceinline int
-  BitSetOffset<A>::max_bit(void) const { return _offset + size() - 1; }
+  BitSetOffset<A>::max_bit(void) const {
+    return _offset + static_cast<int>(size()) - 1;
+  }
 
   template<class A>
   forceinline bool
-  BitSetOffset<A>::valid(int i) const { return _offset <= i && i <= _offset + (int)size() - 1; }
+  BitSetOffset<A>::valid(int i) const {
+    return ((_offset <= i) &&
+            (i <= _offset + static_cast<int>(size()) - 1));
+  }
 
   template <class A, class Char, class Traits>
   std::basic_ostream<Char,Traits>&
